+++ src/scheduler/workers/unpublish/weekly/findListsInAnnualReview.ts
@@ -1,0 +1,19 @@
+import { startOfToday } from "date-fns";
+import { prisma } from "server/models/db/prisma-client";
+
+export async function findListsInAnnualReview() {
+  const today = startOfToday().toISOString();
+  return prisma.list.findMany({
+    where: {
+      nextAnnualReviewStartDate: {
+        lte: today,
+        not: null,
+      },
+      items: {
+        some: {
+          isAnnualReview: true,
+        },
+      },
+    },
+  });
+}
